# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "MovieJoy API",
    "description": "API for managing movies from MovieJoy"
    },
    "paths": {
        "/api/login/{type}": {
            "post": {
                "summary": "Login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type of login",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/api/movies/{cinema}": {
            "get": {
                "summary": "Get all movies from a cinema",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cinema",
                        "in": "path",
                        "description": "Cinema name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Insert token here",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number of the results to retrieve",
                        "required": false,
                        "type": "integer",
                        "default": 1,
                        "format": "int32"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "type": "integer",
                        "default": 4,
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movies",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Movie"
                            }
                        }
                    },
                    "404": {
                        "description": "No movies found"
                    }
                }
            },
            "put": {
                "summary": "Update movies from a cinema",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cinema",
                        "in": "path",
                        "description": "Cinema name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Insert token here",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movies updated successfully"
                    },
                    "500": {
                        "description": "Unexpected error"
                    }
                }
            },
            "delete": {
                "summary": "Delete all records from a cinema",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "cinema",
                        "in": "path",
                        "description": "Cinema name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-Delete-Password",
                        "in": "header",
                        "description": "Password for deleting",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Insert token here",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses":{
                    "200": {
                        "description": "Everything deleted"
                    },
                    "401": {
                        "description": "Incorect password"
                    },
                    "500":{
                        "description": "Invalid input"
                    }
                }
            }
        }
    },
    "definitions": {
        "Movie": {
            "type": "object",
            "properties": {
                "id_local": {
                    "type": "integer",
                    "format": "int64"
                },
                "title": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}